rename linters_forge_yml.py to lint_forge_yml.py
